Fix use of relative paths by using symlinks
This commit fixes the organization of the repository so that it is
still possible to use relative paths inside `kpa_project_dir` and
`marp_output_file` Ansible variables.
This is possible by using the symlinks inside the playbooks folder
pointing to ../projects and ../slides.
It also updates the README accordingly.

Resolve #1