Improve symlink management in output dir
We were overwriting any existing symlink inside the output dir (which is
necessary to marp for its pdf rendering), but this can lead to errors and
changes over local mapped directories, and we don't want it.
This commit checks for an existing projects symlink inside the output directory
and ONLY if not present it creates it (and then delete it).