Use new base container & push directly to quay.io
This commit should solve once and for all the problems related to the container
cache, since with this the build now happens on GitHub Actions.
We now use the base container kpa-marp-pandoc [1] that covers all KPA needed
tools and dependencies, and let us prepare the new container with an improved
speed and lightness.

[1] https://github.com/mmul-it/kpa-marp-pandoc

Fixes #8